Reading timeseries languages directly from controller
