--- pinot-broker/src/main/java/org/apache/pinot/broker/api/resources/PinotBrokerTimeSeriesConfigResource.java
@@ -1,70 +1,0 @@
-/**
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-package org.apache.pinot.broker.api.resources;
-
-import io.swagger.annotations.Api;
-import io.swagger.annotations.ApiKeyAuthDefinition;
-import io.swagger.annotations.Authorization;
-import io.swagger.annotations.SecurityDefinition;
-import io.swagger.annotations.SwaggerDefinition;
-import java.util.Arrays;
-import java.util.List;
-import java.util.stream.Collectors;
-import javax.ws.rs.GET;
-import javax.ws.rs.Path;
-import javax.ws.rs.Produces;
-import javax.ws.rs.core.Application;
-import javax.ws.rs.core.Context;
-import javax.ws.rs.core.MediaType;
-import org.apache.pinot.broker.broker.BrokerAdminApiApplication;
-import org.apache.pinot.spi.env.PinotConfiguration;
-import org.apache.pinot.tsdb.spi.PinotTimeSeriesConfiguration;
-
-import static org.apache.pinot.spi.utils.CommonConstants.SWAGGER_AUTHORIZATION_KEY;
-
-/**
- * Resource to get timeseries language configuration for the broker.
- */
-@Api(tags = "TimeSeriesConfig", authorizations = {@Authorization(value = SWAGGER_AUTHORIZATION_KEY)})
-@SwaggerDefinition(securityDefinition = @SecurityDefinition(apiKeyAuthDefinitions = @ApiKeyAuthDefinition(name =
-    javax.ws.rs.core.HttpHeaders.AUTHORIZATION, in = ApiKeyAuthDefinition.ApiKeyLocation.HEADER,
-    key = SWAGGER_AUTHORIZATION_KEY,
-    description = "The format of the key is  ```\"Basic <token>\" or \"Bearer <token>\"```")))
-@Path("/")
-public class PinotBrokerTimeSeriesConfigResource {
-
-  @Context
-  private Application _application;
-
-  @GET
-  @Path("/timeseries/languages")
-  @Produces(MediaType.APPLICATION_JSON)
-  public List<String> getTimeSeriesLanguages() {
-    PinotConfiguration pinotConfiguration =
-        (PinotConfiguration) _application.getProperties().get(BrokerAdminApiApplication.PINOT_CONFIGURATION);
-
-    String enabledLanguagesConfig = pinotConfiguration.getProperty(
-        PinotTimeSeriesConfiguration.getEnabledLanguagesConfigKey(), "");
-
-    return Arrays.stream(enabledLanguagesConfig.split(","))
-        .map(String::trim)
-        .filter(lang -> !lang.isEmpty())
-        .collect(Collectors.toList());
-  }
-}
